lib: tenstorrent: banner: add TT_BOOT_BANNER_FINGERPRINT
Add TT_BOOT_BANNER_FINGERPRINT, which generates a unique identifier for
every build. If build artifacts for a given application are saved, this
can be used to match a firmware exactly with the build that created it.

Note the fingerprint is only regenerated when cmake is run, so if some
part of the firmware is modified outside of cmake the fingerprint will
not be regenerated.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>