readme: update getting started instructions
The previous getting started guide directed the user to rely on upstream
zephyr's steps for setting up a python virtual environment and
installing the SDK.

However, if these steps are followed to the letter the user will end up
with a separate ~/zephyrproject folder with their virtual environment
installed (and potentially zephyr as well, if they follow the entire
guide).

IMO, it seems reasonable to duplicate portions of the getting started
guide where we deviate from them. Another alternative would be to direct
users to use the getting started guide, but substitute "~/zephyrproject"
out with our module name wherever they see fit.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>