app: pytest: rework pytest harness setup
With tt-kmd 2.4.0, reusing a file descriptor to a previously reset BH
device will cause a -ENODEV error. Since we previously used the same
pyluwen chip object as a session test harness in pytest and reset the
BH chip multiple times in tests we will encounter this error.

To work around this, rework the pytest harness setup to the following
design:
- session harness used to flash the device once at the start of testing
- per-function harness used to detect that the ARC chip is online and
  responsive (and rescan PCIe if it is not)

The per-function harness does not return a handle to the ARC chip, so
each test can freely create and discard handles. This way we can safely
discard handles when the ARC chip resets and avoid hitting the -ENODEV
error.

Additionally, rework the vuart test- we now have a working console
within our twister harness, so we can use that rather than launching
tt-console directly.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>