scripts: tooling: move to new blackhole-recovery script
Move to a new blackhole recovery script. This script no longer uses
sequential recovery steps- instead it directly reprograms the entire
firmware for the SMC and DMC, and rewrites the card UPI.

This style of recovery is more stable, since the DMFW never is permitted
to run the update sequence (which can block the pyocd debugger from
connecting)

The new recovery script uses "recovery bundles", which are similar to
firmware bundles for production releases but contain a build of MCUBoot
for the DMC, and package the EEPROM firmware alongside the firmware
intended for the SMC EEPROM.

This update strategy prevents the DMFW from ever triggered a self-update
via tt-fwupdate, which avoids the connection error described above.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>