zephyr: patches: drop twister rtt patch, since we now have pty scripts
Since we have `scripts/dmc_rtt.py` and `scripts/smc_console.py`, we no
longer require the twister rtt patch (almost).

Also, it was rejected upstream, so this is a good opportunity to shed tech
debt.

However, the `--flash-before` logic is broken in twister upstream (at
least 4.2.0).

Keep this smaller change in-tree until it is resolved upstream.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>