zephyr: patches: update status on patches
As of tt-z-p v18.10.0, we had 18 patches to apply on top of Zephyr. The
goal is to get that as close to zero prior to the v4.3.0 release.

This change updates metadata for a number of patches.

A summary of changes

* 1 of the 18 patches can be removed (some work required)
* 5 of the 18 patches have already been merged upstream
* 4 of the 5 stm32 smbus/i2c patches look upstreamable (PRs shortly)

So that leaves 8 / 18 patches to account for. The following patches
have a path forward:

* zephyr/check-compliance.patch (small but contains a few workarounds)
* zephyr/arc-multilvl-int.patch (needs review comments to be addressed)
* zephyr/gpio-pca-series-pin-get-config.patch (needs some minor updates)
* zephyr/pwm-api-tests-skip-zero-duty-cycle-config.patch
  (upstream testsuite needs to be reworked)
* zephyr/k-busy-wait-ns.patch (needs upstream PR + testsuite)
* zephyr/dirty-hack-for-tt-hwinit.patch
  (can be discarded when drivers have been zephyrized)

And the following patches do not have a clear path forward:

* zephyr/i2c-dw-multi-ints.patch (kind of a hack - can we do better?)
* zephyr/smbus-cancel.patch
  (this needs to be redone taking upstream direction into account)

So only 2 / 18 patches require moderate to significant attention to
mitigate.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>