lib: tenstorrent: log_ringbuf: support direct buffer access
Support direct buffer access within log_ringbuf, so that the ring buffer
memory can be read directly without copying. This also allows us to
attempt to transmit data from the ringbuffer before we actually mark it
as consumed.

This fixes a bug with the DMFW logging implementation where the DMFW
would stream the initial log data via SMBUS before the SMC was ready to
read it. With get/finish claims, we only actually consume the log data
if the SMBUS block write succeeds, so the early portion of logs aren't
lost.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>