ci: use tt-console outright instead of smc console
We are always launching tt-console from smc console, and if we use
tt-console explicitly we no longer need to use --flash-before for smc
e2e tests.

Note we still flash before for the smoke tests, because when tt-console
rescans the bus tt-kmd will set a watchdog timeout. If we then attempt
to load unit test firmware the watchdog would fire and the chip would be
reset.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>