boards: tenstorrent: reduce flash programming frequency
Flash performance test fails intermittently when the programming
frequency is set to 15 MHz, complaining that read data does not
match what is programmed. Testing locally the mismatch seems to be that
large sections of the flash are 0xff when they should be programmed with
sequential bytes, likely indicating the TX FIFO underflowed and CS was
deasserted.

To resolve this, lower the programming frequency of the flash. The smoke
tests have been run locally on a P100a for fifty iterations to validate
this change, and passed each time.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>