scripts: add script to dump SMC state data
Add a script to dump SMC state data via PCIe that may be useful for
debugging. This script can be executed as a module, or in standalone
mode. It requires the ARC chip to be accessible via PCIe, but does not
require the firmware to be in a working state.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>