boards: tt_blackhole: p100: remove max6639 from devicetree
The max6639 fan controller on p100 (scrappy) is hardwired to 100% due to
a workaround for incorrect electrical interfacing between the fan and the
fan controller.

Making the fan controller usable would require a bodge of some kind, which
isn't someting that we want to do at scale, so simply remove the fan
controller, since it's not really useful at this point.

See #595 for more details.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>