app: dmc: mcuboot: always reinit RTT at boot
Always reinit RTT at boot within MCUBoot. Turning this on means we only
will have logs from the current boot of the DMC. This is preferable,
because otherwise multiple boots of the DMC will eventually result in
the RTT log buffer overflowing, and no new logs will be produced until
we clear the buffer by reading it.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>