ci: stop using github.ref
The default `ref` for `actions/checkout` is

```
The branch, tag or SHA to checkout. When checking out the repository that
triggered a workflow, this defaults to the reference or SHA for that
event. Otherwise, uses the default branch.
```

We generally want our CI to use either the tag or SHA, so the default
`ref` for `actions/checkout` is correct.

However, `github.ref` is

```
The fully-formed ref of the branch or tag that triggered the workflow run.
For workflows triggered by push, this is the branch or tag ref that was
pushed. For workflows triggered by pull_request, this is the pull request
merge branch.
```

For push, we do not want the checkout to align with the branch, but the
unique `sha`.

For pull requests, we do want the ref to be equal to
`github.event.pull_request.number`, since subsequent force-pushes on the
same branch should cancel previous jobs with the same PR number.

https://docs.github.com/en/actions/reference/workflows-and-actions\
/contexts#github-context

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>