zephyr: blobs: add mcuboot test trailer binary
Add a "test trailer" binary to the blobs. This blob is a 4KB sector
which can be placed at the end of a mcuboot image section in the
tt-boot-fs. This trailing will mark the image as a candidate for
swapping, which will cause mcuboot to run the image in test mode on the
next boot.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>