app: smc: pytest: modify verbosity and use failure rate
Reduce the verbosity on some of the tests that execute e.g. 1000 times
(although failures are still always logged).

Additionally, use a variable to group failure rate for repeated
tests. For most tests, a single failure is still enough to fail the test
but for some 10 / 1000 is probably a good threshold.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>