Revert "app: smc: pytest: modify verbosity and use failure rate"
This reverts commit 46dfbc45791a8dee2ad4033226de46db6fdb6e09. In my
opinion, we should not be setting thresholds for PCIe enumeration or
reset failures- we should be targeting 100% pass rates for these tests.

There are a few reasons I hold this view:
- Strict pass criteria for these tests lets us catch intermittent errors
  early. If we allow a threshold for errors, then when we do start
  accumulating enough bugs to hit the threshold we may be chasing
  several different bugs at once
- Some of these tests (like PCIe enumeration) should *never* fail.
  Systems should enumerate every boot, full stop. Anything less is a
  firmware issue in my opinion

I'm open to changing my viewpoint here if others disagree, but I would
like to see more discussion than occurred on #654 prior to merge.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>