zephyr: patches: remove pwm_api test patch and update to relevant commit
Remove the pwm_api testsuite patch since it was merged upstream,
introducing test-local Kconfigs for platform specific configuration.

Modify overlay file to use pwm-test alias and conf file to specify testing
on port 1 to validate API. Port 1 is not connected to the fan controller,
so this testsuite will only be used to validate the API, but the actual fan
controller is tested under tests/drivers/tenstorrent/tt_blackhole/max6639.

Signed-off-by: Alexander Lay <alexanderlay@tenstorrent.com>