scripts: dmc_reset + pcie_utils: reduce verbosity
Currently, the dirty reset test logs about 6000 messages for its 1000 test
iterations, and probably close to 5500 of those messages do not serve a
useful purpose.

It's better to log errors with useful information than to log the same
information over and over again for what is just expected behaviour.

This should make it easier to comb through CI logs to determine the root
cause of a failure.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>