zephyr: patches: check-compliance: kconfig: exclude default pwm port
The compliance workflow was failing due to Zephyr's KconfigCheck test not
being able to pick up Kconfig options defined in upstream tests and
samples.

```shell
ERROR   : Test Kconfig failed:
...
CONFIG_DEFAULT_PWM_PORT \
 test-conf/tests/drivers/pwm/pwm_api/boards/\
 tt_blackhole_tt_blackhole_dmc.conf:21
```

The long-term solution for this is to parse `Kconfig*` files that exist
in testsuites and samples in the `check_compliance.py` script upstream.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>