app: smc: ensure bl2 builds sign smc fw and mcuboot uses specified key
Previously, even when the preferred BL2 signature key was used, the
MCUBoot for SMC would still build with the incorrect key, and SMC FW would
still be signed with the incorrect key.

The changes included in this commit as well as the command below will
correctly sign with the key file specified on the commandline.
```shell
west build --sysbuild -p -b tt_blackhole@p100a/tt_blackhole/smc app/smc \
  -- -DFILE_SUFFIX=bl2 \
  -DSB_CONFIG_BL2_SIGNATURE_KEY_FILE="\"$PWD/mykey.pem\"" \
  -DSB_CONFIG_BOOT_SIGNATURE_KEY_FILE="\"$PWD/mykey.pem\""
```

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>