app: dmc_rom_update: erase image magic in start of slot1
If DMC slot1 in the BL2 scheme has an new image in it,
and the old pre-BL2 firmware is running on the DMC, when the DMFW rom
update is attempted the bootstrap process will fail because of the magic
in the start of the DMC slot1. To resolve this, erase slot1 of the new
BL2 image during the rom update process.

Fixes #751

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>