ci: add ci_boards.json file to specify which boards we run in CI [v2]
Add a `ci_boards.json` file, similar to the existing `boards.json` file.

While the `boards.json` file tells us all of the boards we build for, the
`ci_boards.json` file tells us all of the boards that we test in hardware
CI.

Unlike the `boards.json` file, the `ci_boards.json` file is not just a
list of strings, but a list of objects of the form:

```json
{
  "board": "<board-name>",
  "runs-on": "<self-hosted-runner-label>",
  "metal-image": "<container-image-for-metal-test>"
```

With that, the JSON structure will provide all of the requirements of the
`matrix.config` field in `hardware-smoke.yml` and `hardware-long.yml`.

A list of boards suitable for bash scripts can be obtained via

```shell
CI_BOARDS="$(jq -r -c ".[].board" .github/ci_boards.json)"
CI_BOARDS="${CI_BOARDS//$'\n'/ }" # Remove embedded newlines
```

Currently, the same data is maintained in 3 separate locations:

- hardware-long.yml
- hardware-smoke.yml
- docs.yml

So reducing that to 1 source of truth will help with maintainability.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>