Increase SMC to 100KHz ticks/sec and enable tickless
Our shortest period timer is DVFS at 1ms, but to synchronize it with
DMC board power notifications we need to adjust it in units << 1ms. 1%
of 1ms is chosen as the adjustment granularity because DVFS is coded to
assume it's triggered every 1ms. This gives a minimum system tick rate
of 100/ms or 100KHz.

Rather than taking 100,000 timer interrupts per second, enable tickless
so that Zephyr computes how much longer until the next interrupt.

Signed-off-by: Andrew Lewycky <alewycky@tenstorrent.com>