treewide: dts: run all devicetree files through dt-linter
There is a new tool used upstream called dts-linter. It depends on having
NodeJS installed.

Approximate steps to install it on Ubuntu are:

```shell
sudo apt install -y nodejs node-http-proxy-agent node-https-proxy-agent \
  node-chalk
```

A manual run fixing all files in tree was done with this
```shell
/usr/local/bin/dts-linter --formatFixAll \
  $(for i in $(find * -name '*.dts' -o -name '*.dtsi'); do \
  echo -n " --file $i"; done; echo "")
```

If `dt-linter` is available in `$PATH`, then it should be run by the usual
`check_compliance.py` in Zephyr automatically.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>