app: smc: pytest: remove chip object if arc timeout is set by KMD
If arc timeout is set to 10 seconds by kmd, remove the chip object after
we detect it again. Otherwise, we occasionally hit the -ENODEV error
that occurs when trying to reuse a file descriptor from before the chip
was reset.

Testing locally to P300C, this appears to resolve the intermittent error
with the watchdog test

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>