scripts: tt_boot_fs: BootFs: evaluate writes in output methods
This change fixes an issue in `tt_boot_fs.py` that reduced functionality
of the `BootFs` object. Specifically, the `writes` property was finalized
within the constructor and later evaluated in different methods that
convert to various output formats.

Between initialization of the object and when e.g. the `to_binary()`
method is called, contents of various sections may get updated.

However, if the `writes` property is only evaluated within the
constructor, then any changes made to sections are lost.

Therefore, in this change, we convert the `writes` property into a
`writes()` method, and only store the order, entries, and failover within
the constructor itself. The `writes()` medhod is evaluated within the
`to_binary()` and `to_intel_hex()` methods.

This allows for post-constructor modification of the contents of each
entry.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>