scripts: tt_boot_fs: add an option to display checksum in hexdump()
The `hexdump()` function is a convenient way to view contents of various
blobs. However, it can be difficult to know if contents differ when
comparing with a blob that is really similar via manual scan.

Add an argument to tell `hexdump()` to also print the boot fs checksum
after displaying hex contents.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>