drivers: misc: bh_fwtable: move spirom_protobufs from bh_arc to driver
This change moves the `spirom_protobufs` protocol definition files from
`lib/tenstorrent/bh_arc` to `drivers/misc/bh_fwtable`, since the
`bh_fwtable` driver effectively provides an interface for accessing
firmware table data.

It's important to note that we only have one "driver" for firmware table
data at the moment, for Blackhole, and that it isn't abstracting things
for multiple SoCs. Whether or not we _could_ reuse that for other SoCs is
TBD, but that would be kind of ideal in some ways. In that case, it might
make sense to move the firmware table definitions to a common directory
under `soc/tenstorrent`.

For now, it's probably safe to couple the protocol definitions with the
`bh_fwtable` driver.

The generation of firmware table data, on the other hand, is pretty closely tied
to board definitions. So the logic to generate protobuf-encoded binary messages
from per-board definition files is moved to `boards/tenstorrent/tt_blackhole`.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>