scripts: tt-console: remove and rescan using sudo if not root
When manually running `./scripts/ci/run-smoke.sh -t smc p100a -- -c`,
there were permission errors removing and rescanning PCIe devices.

This is because tooling would only work previously if either the
`tt-console` binary was setuid (which is semi-dangerous) or run as
root.

Instead of assuming, check if the effective UID of the process is
0 (i.e. root) if so, proceed as normal. Otherwise, attempt to run
the same effective commands via `sudo`.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>