scripts: requirements.txt: update to power aware Luwen
Update to Luwen which opens file descriptor to TT device
with O_APPEND to keep device in low power state.

Use new luwen command for set_power_state to update the
power state of the device for the tensix_reset test. Do this
in the same commit to maintain bisect-ability.

Signed-off-by: James Growden <jgrowden@tenstorrent.com>